+++ frontend/packages/console-dynamic-plugin-sdk/src/app/components/index.ts
@@ -1,6 +1,9 @@
+export { default as CamelCaseWrap } from './utils/camel-case-wrap';
 export { default as GenericStatus } from './status/GenericStatus';
 export { default as PopoverStatus } from './status/PopoverStatus';
+export { default as StatusComponent } from './status/Status';
 export { default as StatusIconAndText } from './status/StatusIconAndText';
 export { default as ResourceStatus } from './utils/resource-status';
 
 export * from './status/icons';
+export * from './status/statuses';
