+++ frontend/packages/console-dynamic-plugin-sdk/src/app/constants.ts
@@ -1,0 +1,1 @@
+export const DASH = '-';
