+++ frontend/packages/console-shared/src/components/status/statuses.tsx
@@ -1,38 +1,35 @@
 import * as React from 'react';
-import { HourglassHalfIcon, InProgressIcon } from '@patternfly/react-icons';
+import { HourglassHalfIcon } from '@patternfly/react-icons';
 import { useTranslation } from 'react-i18next';
-import { GenericStatus } from '@console/dynamic-plugin-sdk';
 import {
-  RedExclamationCircleIcon,
-  GreenCheckCircleIcon,
-  YellowExclamationTriangleIcon,
-  BlueInfoCircleIcon,
-} from './icons';
-import { StatusComponentProps } from './types';
+  GenericStatus,
+  StatusComponentProps,
+  ErrorStatus as SdkErrorStatus,
+  InfoStatus as SdkInfoStatus,
+  ProgressStatus as SdkProgressStatus,
+  SuccessStatus as SdkSuccessStatus,
+} from '@console/dynamic-plugin-sdk';
+import { YellowExclamationTriangleIcon } from './icons';
 
-export const ErrorStatus: React.FC<StatusComponentProps> = (props) => {
+export const ErrorStatus: React.FC<StatusComponentProps> = ({ title, ...props }) => {
   const { t } = useTranslation();
-  return (
-    <GenericStatus
-      {...props}
-      Icon={RedExclamationCircleIcon}
-      title={props.title || t('console-shared~Error')}
-    />
-  );
+  return <SdkErrorStatus {...props} title={title || t('console-shared~Error')} />;
 };
-ErrorStatus.displayName = 'ErrorStatus';
 
-export const InfoStatus: React.FC<StatusComponentProps> = (props) => {
+export const InfoStatus: React.FC<StatusComponentProps> = ({ title, ...props }) => {
   const { t } = useTranslation();
-  return (
-    <GenericStatus
-      {...props}
-      Icon={BlueInfoCircleIcon}
-      title={props.title || t('console-shared~Information')}
-    />
-  );
+  return <SdkInfoStatus {...props} title={title || t('console-shared~Information')} />;
 };
-InfoStatus.displayName = 'InfoStatus';
+
+export const ProgressStatus: React.FC<StatusComponentProps> = ({ title, ...props }) => {
+  const { t } = useTranslation();
+  return <SdkProgressStatus {...props} title={title || t('console-shared~In progress')} />;
+};
+
+export const SuccessStatus: React.FC<StatusComponentProps> = ({ title, ...props }) => {
+  const { t } = useTranslation();
+  return <SdkSuccessStatus {...props} title={title || t('console-shared~Healthy')} />;
+};
 
 export const PendingStatus: React.FC<StatusComponentProps> = (props) => {
   const { t } = useTranslation();
@@ -45,30 +42,6 @@ export const PendingStatus: React.FC<StatusComponentProps> = (props) => {
   );
 };
 PendingStatus.displayName = 'PendingStatus';
-
-export const ProgressStatus: React.FC<StatusComponentProps> = (props) => {
-  const { t } = useTranslation();
-  return (
-    <GenericStatus
-      {...props}
-      Icon={InProgressIcon}
-      title={props.title || t('console-shared~In progress')}
-    />
-  );
-};
-ProgressStatus.displayName = 'ProgressStatus';
-
-export const SuccessStatus: React.FC<StatusComponentProps> = (props) => {
-  const { t } = useTranslation();
-  return (
-    <GenericStatus
-      {...props}
-      Icon={GreenCheckCircleIcon}
-      title={props.title || t('console-shared~Healthy')}
-    />
-  );
-};
-SuccessStatus.displayName = 'SuccessStatus';
 
 export const WarningStatus: React.FC<StatusComponentProps> = (props) => {
   const { t } = useTranslation();
