--- frontend/packages/console-shared/src/components/status/types.ts
@@ -1,7 +1,0 @@
-export type StatusComponentProps = {
-  title?: string;
-  iconOnly?: boolean;
-  noTooltip?: boolean;
-  className?: string;
-  popoverTitle?: string;
-};
